+++ example/android/app/src/main/kotlin/com/example/webview_plus_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.webview_plus_example
+package me.noam.webview_plus_example
 
 import io.flutter.embedding.android.FlutterActivity