+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Update onLoadStart / onLoadStop / onReceivedError / onNavigationRequest and remove shouldOverrideUrlLoading (now use onNavigationRequest only)
+- Add onDOMContentLoaded / onWindowFocus / onWindowBlur
+
 ## 0.1.0
 
 - Initial Commit